PLPEnum: Initial commit
Enumeration tool used for Hack The Box challenges initial commit.  This is a work in progress with very sharp edges.  I am still in the learning process and working on developing the logic in this tool so it can perform intelligent enumeration and basic attacks, info gathering tasks.

Currently this is tailored to Hack The Box challenges that I am working on.  I will generalize it in the future and port it to python.